+++ Sources/SwiftGen/Commander/ConfigCommands.swift
@@ -9,6 +9,26 @@ import PathKit
 import StencilSwiftKit
 import SwiftGenKit
 
+extension Config {
+  fileprivate func run(cmd: String, verbose: Bool) throws {
+    guard let parserCmd = ParserCLI.command(named: cmd) else {
+      throw Config.Error.missingEntry(key: cmd)
+    }
+
+    for var entry in commands[cmd] ?? [] {
+      entry.makingRelativeTo(inputDir: inputDir, outputDir: outputDir)
+      if verbose {
+        for item in entry.commandLine(forCommand: cmd) {
+          logMessage(.info, " $ \(item)")
+        }
+      }
+
+      try entry.checkPaths()
+      try entry.run(parserCommand: parserCmd)
+    }
+  }
+}
+
 extension ConfigEntryOutput {
   func checkPaths() throws {
     guard self.output.parent().exists else {
@@ -30,9 +50,10 @@ extension ConfigEntry {
   }
 
   func run(parserCommand: ParserCLI) throws {
-    let parser = try parserCommand.parserType.init(options: [:]) { msg, _, _ in
+    let parser = try parserCommand.parserType.init(options: options) { msg, _, _ in
       logMessage(.warning, msg)
     }
+
     let filter = try Filter(pattern: self.filter ?? parserCommand.parserType.defaultFilter)
     try parser.searchAndParse(paths: inputs, filter: filter)
     let context = parser.stencilContext()
@@ -54,58 +75,61 @@ extension ConfigEntry {
 
 // MARK: - Commands
 
-private let configOption = Option<Path>(
-  "config",
-  default: "swiftgen.yml",
-  flag: "c",
-  description: "Path to the configuration file to use",
-  validator: checkPath(type: "config file") { $0.isFile }
-)
-
-// MARK: Lint
-
-let configLintCommand = command(
-  configOption
-) { file in
-  try ErrorPrettifier.execute {
-    logMessage(.info, "Linting \(file)")
-    let config = try Config(file: file)
-    config.lint()
+enum ConfigCLI {
+  private enum CLIOption {
+    static let config = Option<Path>(
+      "config",
+      default: "swiftgen.yml",
+      flag: "c",
+      description: "Path to the configuration file to use",
+      validator: checkPath(type: "config file") { $0.isFile }
+    )
   }
-}
 
-// MARK: Run
+  // MARK: Lint
 
-let configRunCommand = command(
-  configOption,
-  Flag("verbose", default: false, flag: "v", description: "Print each command being executed")
-) { file, verbose in
-  do {
+  static let lint = command(
+    CLIOption.config
+  ) { file in
     try ErrorPrettifier.execute {
+      logMessage(.info, "Linting \(file)")
       let config = try Config(file: file)
+      config.lint()
+    }
+  }
 
-      if verbose {
-        logMessage(.info, "Executing configuration file \(file)")
-      }
-      try file.parent().chdir {
-        for (cmd, entries) in config.commands {
-          for var entry in entries {
-            guard let parserCmd = allParserCommands.first(where: { $0.name == cmd }) else {
-              throw Config.Error.missingEntry(key: cmd)
-            }
-            entry.makingRelativeTo(inputDir: config.inputDir, outputDir: config.outputDir)
-            if verbose {
-              for item in entry.commandLine(forCommand: cmd) {
-                logMessage(.info, " $ \(item)")
-              }
-            }
-            try entry.checkPaths()
-            try entry.run(parserCommand: parserCmd)
+  // MARK: Run
+
+  static let run = command(
+    CLIOption.config,
+    Flag("verbose", default: false, flag: "v", description: "Print each command being executed")
+  ) { file, verbose in
+    do {
+      try ErrorPrettifier.execute {
+        let config = try Config(file: file)
+
+        if verbose {
+          logMessage(.info, "Executing configuration file \(file)")
+        }
+        try file.parent().chdir {
+          for cmd in config.commands.keys.sorted() {
+            try config.run(cmd: cmd, verbose: verbose)
           }
         }
       }
+    } catch let error as Config.Error {
+      logMessage(.error, error)
+      logMessage(
+        .error,
+        """
+        It seems like there was an error running SwiftGen. Please verify that your configuration file is valid by \
+        running:
+        > swiftgen config lint
+
+        If you have any other questions or issues, we have extensive documentation and an issue tracker on GitHub:
+        > https://github.com/SwiftGen/SwiftGen
+        """
+      )
     }
-  } catch let error as Config.Error {
-    logMessage(.error, error)
   }
 }