+++ Tests/Fixtures/CompilationEnvironment/IB-macOS-NamespacedLogin.swift
@@ -1,0 +1,6 @@
+import Cocoa
+
+enum ExtraModule {
+  final class LoginSegue: NSStoryboardSegue {}
+  final class LoginViewController: NSWindowController {}
+}