+++ Tests/SwiftGenKitTests/ColorsJSONFileTests.swift
@@ -10,8 +10,9 @@ import XCTest
 
 class ColorsJSONFileTests: XCTestCase {
   func testFileWithDefaults() throws {
-    let parser = Colors.Parser()
-    parser.palettes = [try Colors.JSONFileParser().parseFile(at: Fixtures.path(for: "colors.json", sub: .colors))]
+    let parser = try Colors.Parser()
+    let jsonParser = Colors.JSONFileParser(options: try ParserOptionValues(options: [:], available: []))
+    parser.palettes = [try jsonParser.parseFile(at: Fixtures.path(for: "colors.json", sub: .colors))]
 
     let result = parser.stencilContext()
     XCTDiffContexts(result, expected: "defaults", sub: .colors)
@@ -19,7 +20,8 @@ class ColorsJSONFileTests: XCTestCase {
 
   func testFileWithBadSyntax() {
     do {
-      _ = try Colors.JSONFileParser().parseFile(at: Fixtures.path(for: "bad-syntax.json", sub: .colors))
+      let jsonParser = Colors.JSONFileParser(options: try ParserOptionValues(options: [:], available: []))
+      _ = try jsonParser.parseFile(at: Fixtures.path(for: "bad-syntax.json", sub: .colors))
       XCTFail("Code did parse file successfully while it was expected to fail for bad syntax")
     } catch Colors.ParserError.invalidFile {
       // That's the expected exception we want to happen
@@ -30,7 +32,8 @@ class ColorsJSONFileTests: XCTestCase {
 
   func testFileWithBadValue() {
     do {
-      _ = try Colors.JSONFileParser().parseFile(at: Fixtures.path(for: "bad-value.json", sub: .colors))
+      let jsonParser = Colors.JSONFileParser(options: try ParserOptionValues(options: [:], available: []))
+      _ = try jsonParser.parseFile(at: Fixtures.path(for: "bad-value.json", sub: .colors))
       XCTFail("Code did parse file successfully while it was expected to fail for bad value")
     } catch Colors.ParserError.invalidHexColor(path: _, string: "this isn't a color", key: "ArticleTitle"?) {
       // That's the expected exception we want to happen