+++ Tests/TemplatesTests/StringsTests.swift
@@ -24,7 +24,25 @@ class StringsTests: XCTestCase {
       (context: try StencilContext.enrich(context: context, parameters: ["noComments"]),
        suffix: "-noComments"),
       (context: try StencilContext.enrich(context: context, parameters: ["publicAccess"]),
-       suffix: "-publicAccess")
+       suffix: "-publicAccess"),
+      (context: try StencilContext.enrich(
+        context: context,
+        parameters: ["localizeFunction=XCTLocFunc"]
+       ),
+       suffix: "-localizeFunction")
+    ]
+  }
+
+  let variationsObjC: VariationGenerator = { name, context in
+    guard name == "localizable" else { return [(context: context, suffix: "")] }
+
+    return [
+      (context: context,
+       suffix: ""),
+      (context: try StencilContext.enrich(context: context, parameters: ["noComments"]),
+       suffix: "-no-comments"),
+      (context: try StencilContext.enrich(context: context, parameters: ["headerName=headerName-from-parameter.h"]),
+       suffix: "-headerName")
     ]
   }
 
@@ -61,6 +79,26 @@ class StringsTests: XCTestCase {
       contextNames: Contexts.all,
       directory: .strings,
       contextVariations: variations
+    )
+  }
+
+  func testObjectiveCHeader() {
+    test(
+      template: "objc-h",
+      contextNames: Contexts.all,
+      directory: .strings,
+      contextVariations: variationsObjC,
+      outputExtension: "h"
+    )
+  }
+
+  func testObjectiveCSource() {
+    test(
+      template: "objc-m",
+      contextNames: Contexts.all,
+      directory: .strings,
+      contextVariations: variationsObjC,
+      outputExtension: "m"
     )
   }
 }