+++ Tests/TemplatesTests/TestsHelper.swift
@@ -119,8 +119,8 @@ class Fixtures {
     return string(for: name, subDirectory: "templates/\(sub.rawValue.lowercased())")
   }
 
-  static func output(for name: String, sub: Directory) -> String {
-    return string(for: name, subDirectory: "Generated/\(sub.rawValue)")
+  static func output(template: String, variation: String, sub: Directory) -> String {
+    return string(for: variation, subDirectory: "Generated/\(sub.rawValue)/\(template)")
   }
 
   private static func string(for name: String, subDirectory: String) -> String {
@@ -147,7 +147,7 @@ extension XCTestCase {
    
    - Parameter template: The name of the template (without the `stencil` extension)
    - Parameter contextNames: A list of context names (without the `plist` extension)
-   - Parameter directory: The directory to look for files in (correspons to de command)
+   - Parameter directory: The directory to look for files in (corresponds to the command)
    - Parameter resourceDirectory: The directory to look for files in (corresponds to the command)
    - Parameter contextVariations: Optional closure to generate context variations.
    */
@@ -159,7 +159,8 @@ extension XCTestCase {
     outputDirectory: Fixtures.Directory? = nil,
     file: StaticString = #file,
     line: UInt = #line,
-    contextVariations: VariationGenerator? = nil
+    contextVariations: VariationGenerator? = nil,
+    outputExtension: String = "swift"
   ) {
     let templateString = Fixtures.template(for: "\(templateName).stencil", sub: directory)
     let template = StencilSwiftTemplate(
@@ -182,7 +183,7 @@ extension XCTestCase {
       }
 
       for (index, (context: context, suffix: suffix)) in variations.enumerated() {
-        let outputFile = "\(templateName)-context-\(contextName)\(suffix).swift"
+        let outputFile = "\(contextName)\(suffix).\(outputExtension)"
         if variations.count > 1 { print(" - Variation #\(index)... (expecting: \(outputFile))") }
 
         let result: String
@@ -194,14 +195,15 @@ extension XCTestCase {
 
         // check if we should generate or not
         if ProcessInfo().environment["GENERATE_OUTPUT"] == "YES" {
-          let target = Path(#file).parent().parent() + "Fixtures/Generated" + outputDir.rawValue + outputFile
+          let target = Path(#file).parent().parent() + "Fixtures/Generated" + outputDir.rawValue +
+            templateName + outputFile
           do {
             try target.write(result)
           } catch {
             fatalError("Unable to write output file \(target)")
           }
         } else {
-          let expected = Fixtures.output(for: outputFile, sub: outputDir)
+          let expected = Fixtures.output(template: templateName, variation: outputFile, sub: outputDir)
           XCTDiffStrings(result, expected, file: file, line: line)
         }
       }